rythmhacks/backend/requirements.txt
# MediLytica Backend Requirements
# ML Visualization and Collaboration Platform

# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.17

# Machine Learning
scikit-learn==1.5.2
numpy==2.1.3
pandas==2.2.3
# shap==0.45.0  # Model interpretability - requires C++ build tools
umap-learn==0.5.5  # Dimensionality reduction for data visualization

# PyTorch (required for .pt/.pkl PyTorch models)
torch>=2.6.0
# torchvision>=0.20.1  # Uncomment if using vision models

# Optional: TensorFlow/Keras (uncomment if using deep learning)
# tensorflow==2.18.0

# Database
sqlalchemy==2.0.35

# Data Validation
pydantic==2.10.2
pydantic-settings==2.6.1

# WebSocket Support
websockets==14.1

# Utilities
python-dateutil==2.9.0
python-dotenv==1.0.1

# File Handling
openpyxl==3.1.5  # Excel support

# Code Execution
RestrictedPython==7.4  # Secure Python code execution sandbox
